+++ abci/did/v1/deliverTx.go
@@ -166,6 +166,8 @@ func (app *DIDApplication) callDeliverTx(name string, param string, nodeID strin
 		return app.updateNodeProxyNode(param, nodeID)
 	case "RemoveNodeFromProxyNode":
 		return app.removeNodeFromProxyNode(param, nodeID)
+	case "RevokeAccessorMethod":
+		return app.revokeAccessorMethod(param, nodeID)
 	default:
 		return types.ResponseDeliverTx{Code: code.UnknownMethod, Log: "Unknown method name"}
 	}